+++ starsky-tools/localtunnel/localtunnel.js
@@ -11,17 +11,16 @@ const https = require('https');
 // all urls to proxy
 app.all("/*", function (req, res, next) {
 
-  if(req.originalUrl.startsWith("/sockjs-node")) {
+  if (req.originalUrl.startsWith("/sockjs-node")) {
     res.status(503);
     res.json("not allowed");
   }
 
   if (req.originalUrl.startsWith("/api") ||
-  req.originalUrl.startsWith("/account")  ||
-  req.originalUrl.startsWith("/sync/")  ||
-  req.originalUrl.startsWith("/suggest/")  ||
-  req.originalUrl.startsWith("/export/")  )
-  {
+    req.originalUrl.startsWith("/account") ||
+    req.originalUrl.startsWith("/sync/") ||
+    req.originalUrl.startsWith("/suggest/") ||
+    req.originalUrl.startsWith("/export/")) {
     NetCoreAppRouteRoute(req, res, next);
   }
   else {
@@ -118,9 +117,15 @@ console.log("http://localhost:" + port);
 const localtunnel = require('localtunnel');
 
 (async () => {
+
+  // lt -p 8080 -h http://localtunnel.me --local-https false
   const tunnel = await localtunnel({
     subdomain: process.env.SUBDOMAIN,
-    port: port
+    host: 'http://localtunnel.me',
+    port: port,
+    local_https: false
+  }).catch(err => {
+    throw err;
   });
 
   // the assigned public url for your tunnel
@@ -133,6 +138,6 @@ const localtunnel = require('localtunnel');
   });
 
   tunnel.on('close', () => {
-    // tunnels are closed
+    console.log('tunnels are closed');
   });
 })();