+++ starsky/starsky/clientapp/src/components/item-list-view.tsx
@@ -1,5 +1,7 @@
-import React, { memo } from 'react';
+import React, { memo, useEffect } from 'react';
+import useLocation from '../hooks/use-location';
 import { IFileIndexItem } from '../interfaces/IFileIndexItem';
+import { INavigateState } from '../interfaces/INavigateState';
 import ListImageBox from './list-image-box';
 
 interface ItemListProps {
@@ -11,20 +13,37 @@ interface ItemListProps {
  */
 const ItemListView: React.FunctionComponent<ItemListProps> = memo((props) => {
 
-  // todo: implement feature that saves the scroll height
-  // var history = useLocation();
-  // useEffect(() => {
-  //   var state = history.location.state as INavigateState;
-  //   if (!state.fileName) return;
+  // feature that saves the scroll height
+  var history = useLocation();
+  const folderRef = React.useRef<HTMLDivElement>(null);
 
-  //   console.log(state.fileName);
-  // }, [history.location.state]);
+  useEffect(() => {
+    var navigationState = history.location.state as INavigateState;
+
+    if (!navigationState) return;
+    if (!navigationState.filePath) return;
+
+    // for the DOM delay
+    setTimeout(() => {
+      var query = '[data-filepath="' + navigationState.filePath + '"]';
+      var elementList = document.querySelectorAll(query);
+      if (elementList.length !== 1) return;
+
+      window.scrollTo({
+        top: elementList[0] ? (elementList[0] as HTMLDivElement).offsetTop : 0
+      });
+
+      // reset afterwards (when you refresh the state isn't cleared)
+      history.navigate(history.location.href, { replace: true });
+    }, 100);
+
+  }, [history.location.state]);
 
   let items = props.fileIndexItems;
   if (!items) return (<div className="folder">no content</div>);
 
   return (
-    <div className="folder">
+    <div className="folder" ref={folderRef} >
       {items.length === 0 ? props.colorClassUsage.length >= 1 ? <div className="warning-box warning-box--left">Er zijn meer items, maar deze vallen buiten je filters</div> : <div className="warning-box"> Er zijn geen foto's in deze map</div> : ""}
       {
         items.map((item, index) => (
@@ -32,26 +51,6 @@ const ItemListView: React.FunctionComponent<ItemListProps> = memo((props) => {
         ))
       }
     </div>)
-
-
-  // return (
-  //   <>
-  //     <ArchiveContextConsumer>
-  //       {appContext =>
-  //         appContext && (
-  //           <div className="folder">
-  //             {appContext.state.fileIndexItems.length === 0 ? appContext.state.colorClassUsage.length >= 1 ? <div className="warning-box warning-box--left">Er zijn meer items, maar deze vallen buiten je filters</div> : <div className="warning-box"> Er zijn geen foto's in deze map</div> : ""}
-  //             {
-  //               appContext.state.fileIndexItems.map((item, index) => (
-  //                 <ListImageBox item={item} key={item.fileName}></ListImageBox>
-  //               ))
-  //             }
-  //           </div>
-  //         )
-  //       }
-  //     </ArchiveContextConsumer>
-  //   </>
-  // );
 });
 
 export default ItemListView