+++ starsky/starsky/clientapp/src/containers/archive.tsx
@@ -26,15 +26,15 @@ function Archive(archive: IArchiveProps) {
     <>
       <MenuArchive />
       <div className={!sidebar ? "archive" : "archive collapsed"}>
-        {sidebar ? <ArchiveSidebar {...archive}/> : ""}
+        <ArchiveSidebar {...archive} />
 
         <div className="content">
-          <Breadcrumb breadcrumb={archive.breadcrumb} subPath={archive.subPath}/>
-          <RelativeLink relativeObjects={archive.relativeObjects}/>
+          <Breadcrumb breadcrumb={archive.breadcrumb} subPath={archive.subPath} />
+          <RelativeLink relativeObjects={archive.relativeObjects} />
 
-          <ColorClassFilter itemsCount={archive.collectionsCount} subPath={archive.subPath} 
-                            colorClassFilterList={archive.colorClassFilterList} 
-                            colorClassUsage={archive.colorClassUsage}/>
+          <ColorClassFilter itemsCount={archive.collectionsCount} subPath={archive.subPath}
+            colorClassFilterList={archive.colorClassFilterList}
+            colorClassUsage={archive.colorClassUsage} />
           <ItemListView {...archive} colorClassUsage={archive.colorClassUsage}> </ItemListView>
         </div>
       </div>