+++ starsky/starsky/clientapp/src/containers/detailview.tsx
@@ -23,13 +23,10 @@ const DetailView: React.FC<IDetailView> = () => {
 
   // if there is no state
   if (!state) {
-    state = { fileIndexItem: { fileHash: '' }, ...newDetailView() };
+    state = { relativeObjects: newIRelativeObjects(), fileIndexItem: { fileHash: '' }, ...newDetailView() };
   }
 
-  let relativeObjects = newIRelativeObjects();
-  if (state && state.relativeObjects) {
-    relativeObjects = state.relativeObjects;
-  }
+  const [relativeObjects, setRelativeObjects] = React.useState(state.relativeObjects);
 
   const [isDetails, setDetails] = React.useState(new URLPath().StringToIUrl(history.location.search).details);
   useEffect(() => {
@@ -55,15 +52,33 @@ const DetailView: React.FC<IDetailView> = () => {
       if (!state.fileIndexItem.orientation) return;
       if (result.statusCode === 202) {
         // result from API is: "Thumbnail is not ready yet"
-        console.log(state.fileIndexItem.orientation);
         setTranslateRotation(state.fileIndexItem.orientation);
       }
       else if (result.statusCode === 200) {
         // thumbnail is alreay rotated (but need to be called due change of image)
         setTranslateRotation(Orientation.Horizontal);
       }
+    }).catch((e) => {
+      console.log(e);
     });
   }, [state.fileIndexItem.fileHash]);
+
+  // know if you searching ?t= in url
+  const [isSearchQuery, setIsSearchQuery] = React.useState(!!new URLPath().StringToIUrl(history.location.search).t);
+  useEffect(() => {
+    setIsSearchQuery(!!new URLPath().StringToIUrl(history.location.search).t);
+  }, [history.location.search]);
+
+  // update relative next prev buttons for search queries
+  useEffect(() => {
+    if (state.subPath === "/" || !isSearchQuery) return;
+    FetchGet(new UrlQuery().UrlSearchRelativeApi(state.subPath, new URLPath().StringToIUrl(history.location.search).t, new URLPath().StringToIUrl(history.location.search).p)).then((result) => {
+      if (result.statusCode !== 200) return;
+      setRelativeObjects(result.data);
+    }).catch((err) => {
+      console.log(err);
+    });
+  }, [isSearchQuery, state.subPath]);
 
   useKeyboardEvent(/ArrowLeft/, (event: KeyboardEvent) => {
     if (new Keyboard().isInForm(event)) return;
@@ -86,7 +101,7 @@ const DetailView: React.FC<IDetailView> = () => {
 
     history.navigate(parentDirectory, {
       state: {
-        fileName: state.fileIndexItem.fileName
+        filePath: state.fileIndexItem.filePath
       } as INavigateState
     });
   }, [state.fileIndexItem]);