+++ starsky/starsky/clientapp/src/shared/url-query.ts
@@ -9,26 +9,32 @@ export class UrlQuery {
     return output.replace(/\+/ig, "%2B");
   }
 
-  public UrlLogin() {
+  public UrlLogin(): string {
     return "/account/login?json=true";
   }
 
-  public UrlQuerySearchApi = (query: string, pageNumber = 0) => {
+  public UrlSearchRelativeApi = (f: string, t: string | undefined, pageNumber = 0): string => {
+    return "/api/search/relativeObjects?f=" + f + "&t=" +
+      t +
+      "&p=" + pageNumber;
+  }
+
+  public UrlQuerySearchApi = (query: string, pageNumber = 0): string => {
     return "/api/search?json=true&t=" + query + "&p=" + pageNumber;
   }
 
-  public UrlSearchTrashApi = (pageNumber = 0) => {
+  public UrlSearchTrashApi = (pageNumber = 0): string => {
     return "/api/search/trash?p=" + pageNumber;
   }
 
-  public UrlAccountStatus = () => {
+  public UrlAccountStatus = (): string => {
     return "/account/status";
   }
 
   /**
    * Used with localisation hash
    */
-  public UrlQueryServerApi = (historyLocationHash: string) => {
+  public UrlQueryServerApi = (historyLocationHash: string): string => {
     var requested = new URLPath().StringToIUrl(historyLocationHash);
 
     var urlObject = newIUrl();
@@ -51,7 +57,7 @@ export class UrlQuery {
   /**
    * Get Direct api/index with IUrl
    */
-  public UrlIndexServerApi = (urlObject: IUrl) => {
+  public UrlIndexServerApi = (urlObject: IUrl): string => {
     return "/api/index" + new URLPath().IUrlToString(urlObject)
   }
 
@@ -68,41 +74,41 @@ export class UrlQuery {
   /**
    * POST to this to update meta information
    */
-  public UrlUpdateApi = () => {
+  public UrlUpdateApi = (): string => {
     return "/api/update";
   }
 
   /**
    * POST to this to search and replace meta information like: tags, descriptions and titles
    */
-  public UrlReplaceApi = () => {
+  public UrlReplaceApi = (): string => {
     return "/api/replace";
   }
 
-  public UrlThumbnailImage = (fileHash: string) => {
+  public UrlThumbnailImage = (fileHash: string): string => {
     return "/api/thumbnail/" + fileHash + ".jpg?issingleitem=true";
   }
 
-  public UrlThumbnailJsonApi = (fileHash: string) => {
+  public UrlThumbnailJsonApi = (fileHash: string): string => {
     return "/api/thumbnail/" + fileHash + "?json=true";
   }
 
   // http://localhost:5000/api/downloadPhoto?f=%2F__starsky%2F0001-readonly%2F4.jpg&isThumbnail=True
-  public UrlDownloadPhotoApi = (f: string, isThumbnail: boolean = true) => {
+  public UrlDownloadPhotoApi = (f: string, isThumbnail: boolean = true): string => {
     return "/api/downloadPhoto?f=" + f + "&isThumbnail=" + isThumbnail;
   }
 
   /**
    * url create a zip
    */
-  public UrlExportPostZipApi = () => {
+  public UrlExportPostZipApi = (): string => {
     return "/export/createZip/"
   }
 
   /**
    * export/zip/SR497519527.zip?json=true
    */
-  public UrlExportZipApi = (createZipId: string, json: boolean = true) => {
+  public UrlExportZipApi = (createZipId: string, json: boolean = true): string => {
     return "/export/zip/" + createZipId + ".zip?json=" + json;
   }